fix: US-004 - Extract agent memory facade into agent-memory.ts
Move AgentMemoryClient class from client.ts to agent-memory.ts.
Re-export from client.ts preserves public API via index.ts chain.
client.ts reduced from 413 to 385 NLOC. 173 tests pass.
