+++ sdks/typescript/src/client.ts
@@ -28,13 +28,14 @@ import type {
   CollectionStatsResponse,
   CollectionConfigResponse,
   AgentMemoryConfig,
-  SemanticEntry,
-  EpisodicEvent,
-  ProceduralPattern,
 } from './types';
 import { ValidationError } from './types';
 import { WasmBackend } from './backends/wasm';
 import { RestBackend } from './backends/rest';
+import { AgentMemoryClient } from './agent-memory';
+
+// Re-export for backward compatibility
+export { AgentMemoryClient } from './agent-memory';
 
 /**
  * VelesDB Client
@@ -846,50 +847,5 @@ export class VelesDB {
   agentMemory(config?: AgentMemoryConfig): AgentMemoryClient {
     this.ensureInitialized();
     return new AgentMemoryClient(this.backend, config);
-  }
-}
-
-/**
- * Agent Memory client for semantic, episodic, and procedural memory
- */
-export class AgentMemoryClient {
-  constructor(
-    private readonly backend: IVelesDBBackend,
-    private readonly config?: AgentMemoryConfig
-  ) {}
-
-  /** Configured embedding dimension (default: 384) */
-  get dimension(): number {
-    return this.config?.dimension ?? 384;
-  }
-
-  /** Store a semantic fact */
-  async storeFact(collection: string, entry: SemanticEntry): Promise<void> {
-    return this.backend.storeSemanticFact(collection, entry);
-  }
-
-  /** Search semantic memory */
-  async searchFacts(collection: string, embedding: number[], k = 5): Promise<SearchResult[]> {
-    return this.backend.searchSemanticMemory(collection, embedding, k);
-  }
-
-  /** Record an episodic event */
-  async recordEvent(collection: string, event: EpisodicEvent): Promise<void> {
-    return this.backend.recordEpisodicEvent(collection, event);
-  }
-
-  /** Recall episodic events */
-  async recallEvents(collection: string, embedding: number[], k = 5): Promise<SearchResult[]> {
-    return this.backend.recallEpisodicEvents(collection, embedding, k);
-  }
-
-  /** Store a procedural pattern */
-  async learnProcedure(collection: string, pattern: ProceduralPattern): Promise<void> {
-    return this.backend.storeProceduralPattern(collection, pattern);
-  }
-
-  /** Match procedural patterns */
-  async recallProcedures(collection: string, embedding: number[], k = 5): Promise<SearchResult[]> {
-    return this.backend.matchProceduralPatterns(collection, embedding, k);
   }
 }
